modularized and split concerns into separate dynamic functions
